requirements.txt
requests~=2.24.0
python-dotenv~=0.14.0
python-telegram-bot~=12.4.2